Rename connection to messaging.
