--- docs/reference/connection.md
@@ -1,3 +1,0 @@
-:::async_kernel.connection
-:::async_kernel.connection.base
-:::async_kernel.connection.zmq
