--- src/async_kernel/connection/__init__.py
@@ -1,3 +1,0 @@
-from async_kernel.connection.base import LocalClient
-
-__all__ = ["LocalClient"]
