+++ src/async_kernel/messaging/__init__.py
@@ -1,0 +1,3 @@
+from async_kernel.messaging.base import LocalClient
+
+__all__ = ["LocalClient"]
