+++ tests/conftest.py
@@ -8,9 +8,9 @@ import pytest
 
 import async_kernel
 from async_kernel import Caller, Kernel
-from async_kernel.connection.base import LocalClient
-from async_kernel.connection.zmq import ZMQClient, ZMQConnection
 from async_kernel.interface import Interface
+from async_kernel.messaging.base import LocalClient
+from async_kernel.messaging.zmq import ZMQClient, ZMQConnection
 from async_kernel.typing import Backend, Channel, ExecuteContent, Job, Message, MessageProtocol, MsgHeader, MsgType
 
 if TYPE_CHECKING:
