+++ tests/test_enter_kernel.py
@@ -4,8 +4,8 @@ from typing import TYPE_CHECKING
 
 import pytest
 
-from async_kernel.connection.base import LocalClient
 from async_kernel.interface import Interface
+from async_kernel.messaging.base import LocalClient
 
 if TYPE_CHECKING:
     from async_kernel import Caller
